Refactor executive summary exports for Next.js compliance
